+++ chessAI/chessAI-master/neuralnet.h
@@ -1,11 +1,11 @@
 #ifndef NEURALNET_H
 #define NEURALNET_H
 
-double reluActivation(double x, int flag);
-double tanhActivation(double x, int flag);
-void reluArray(double *A, int m, int n, int flag);
-void tanhArray(double *A, int m, int n, int flag);
-void nadamArray( double *W, double *V, double *M, double *G, int m, int n, int t);
-void nadam(double *w, double *v, double *m, double *g, int t);
+long double reluActivation(long double x, int flag);
+long double tanhActivation(long double x, int flag);
+void reluArray(long double *A, int m, int n, int flag);
+void tanhArray(long double *A, int m, int n, int flag);
+void nadamArray( long double *W, long double *V, long double *M, long double *G, int m, int n, int t);
+void nadam(long double *w, long double *v, long double *m, long double *g, int t);
 
 #endif
